# backend/docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Gift Croudfunding
  description: >-
    Gift croudfunding service API

    Some useful links:

    - [backend repository](https://github.com/gvozdenkov/kupipodariday-backend)

    - [frontend repository](https://github.com/yandex-praktikum/kupipodariday-frontend)

    - [The source API
    definition](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  contact:
    email: gvozdenkov@gmail.com
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Locale dev server
  - url: https://virtserver.swaggerhub.com/zlocate/KupiPodariDay/1.0.0
    description: SwaggerHub API Auto Mocking
...
paths:
  /users/me:
    get:
      tags:
        - users
      operationId: UsersController_findOwn
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponseDto'
      security:
        - bearerAuth: []
    patch:
      tags:
        - users
      operationId: UsersController_update
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponseDto'
        '400':
          description: Ошибка валидации переданных значений
      security:
        - bearerAuth: []
  /users/me/wishes:
    get:
      tags:
        - users
      operationId: UsersController_getOwnWishes
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MeWishFeed'
      security:
        - bearerAuth: []
  /users/{username}:
    get:
      tags:
        - users
      operationId: UsersController_findOne
      parameters:
        - name: username
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublicProfileResponseDto'
      security:
        - bearerAuth: []
  /users/{username}/wishes:
    get:
      tags:
        - users
      operationId: UsersController_getWishes
      parameters:
        - name: username
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserWishesDto'
      security:
        - bearerAuth: []
  /users/find:
    get:
      tags:
        - users
      operationId: UsersController_findMany
      parameters:
        - name: username
          in: query
          required: false
          schema:
            type: string
        - name: email
          in: query
          required: false
          schema:
            type: email
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfileResponseDto'
      security:
        - bearerAuth: []
...
components:
  schemas:
    UserProfileResponseDto:
      required:
        - about
        - avatar
        - createdAt
        - email
        - id
        - updatedAt
        - username
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          maxLength: 64
          minLength: 1
          type: string
          example: user
        about:
          maxLength: 200
          minLength: 1
          type: string
          example: Пока ничего не рассказал о себе
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        email:
          type: string
          example: user@yandex.ru
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateUserDto:
      type: object
      properties:
        username:
          maxLength: 64
          minLength: 1
          type: string
          example: exampleuser
        about:
          maxLength: 200
          minLength: 0
          type: string
          example: exampleuser
        avatar:
          type: string
          example: https://i.pravatar.cc/150?img=3
        email:
          type: string
          example: user@yandex.ru
        password:
          minLength: 2
          type: string
          example: somestrongpassword
...
    MeWishFeed:
      required:
        - copied
        - createdAt
        - description
        - id
        - image
        - link
        - name
        - offers
        - owner
        - price
        - raised
        - updatedAt
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          maxLength: 250
          minLength: 1
          type: string
        link:
          type: string
        image:
          type: string
        price:
          minimum: 1
          type: number
        raised:
          minimum: 1
          type: number
        copied:
          type: number
        description:
          maxLength: 1024
          minLength: 1
          type: string
...
    UserPublicProfileResponseDto:
      required:
        - about
        - avatar
        - createdAt
        - id
        - updatedAt
        - username
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          maxLength: 64
          minLength: 1
          type: string
          example: user
        about:
          maxLength: 200
          minLength: 1
          type: string
          example: Пока ничего не рассказал о себе
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserWishesDto:
      required:
        - copied
        - createdAt
        - description
        - id
        - image
        - link
        - name
        - offers
        - price
        - raised
        - updatedAt
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          maxLength: 250
          minLength: 1
          type: string
        link:
          type: string
        image:
          type: string
        price:
          minimum: 1
          type: number
        raised:
          minimum: 1
          type: number
        copied:
          type: number
        description:
          maxLength: 1024
          minLength: 1
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT